# internal/swagger/spec.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "durpdeploy REST API.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/deployments": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "summary": "List all deployments across all projects.",
        "operationId": "listAllDeployments",
        "responses": {
          "200": {
            "description": "DeploymentListResponse is the paginated envelope for ListDeployments.",
            "schema": {
              "$ref": "#/definitions/DeploymentListResponse"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deployments/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "summary": "Get a deployment.",
        "operationId": "getDeployment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment represents a release executing against an environment.",
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deployments/{id}/approve": {
      "post": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "operationId": "approveDeployment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment represents a release executing against an environment.",
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deployments/{id}/cancel": {
      "post": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "summary": "Cancel a running deployment.",
        "operationId": "cancelDeployment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment represents a release executing against an environment.",
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deployments/{id}/events": {
      "get": {
        "produces": [
          "text/event-stream"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "summary": "Stream deployment events as Server-Sent Events.",
        "operationId": "deploymentEvents",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "StreamResponse is a server-sent event stream.",
            "schema": {
              "$ref": "#/definitions/StreamResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deployments/{id}/logs": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "summary": "List deployment step logs.",
        "operationId": "listDeploymentLogs",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "DeploymentLogListResponse is a list of deployment log entries.",
            "schema": {
              "$ref": "#/definitions/DeploymentLogListResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/deployments/{id}/redeploy": {
      "post": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "operationId": "redeployDeployment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Deployment represents a release executing against an environment.",
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "409": {
            "description": "Conflict error response.",
            "schema": {
              "$ref": "#/definitions/ConflictError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deployments/{id}/retry": {
      "post": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "summary": "Retry a failed or cancelled deployment.",
        "operationId": "retryDeployment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment represents a release executing against an environment.",
            "schema": {
              "$ref": "#/definitions/Deployment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/deployments/{id}/status": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "deployments"
        ],
        "operationId": "getDeploymentStatus",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "DeploymentStatusResponse is the status payload for GetDeploymentStatus.",
            "schema": {
              "$ref": "#/definitions/DeploymentStatusResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/projects/{id}/releases/{relId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "releases"
        ],
        "summary": "Get a release with its variable snapshots.",
        "operationId": "getRelease",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "RelID",
            "description": "The release ID.",
            "name": "relId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ReleaseWithVariablesResponse is the detailed release payload.",
            "schema": {
              "$ref": "#/definitions/ReleaseWithVariablesResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Deployment": {
      "type": "object",
      "title": "Deployment represents a release executing against an environment.",
      "properties": {
        "created_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "environment_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "EnvironmentID"
        },
        "finished_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "FinishedAt"
        },
        "forced": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Forced"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "note": {
          "type": "string",
          "x-go-name": "Note"
        },
        "release_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ReleaseID"
        },
        "started_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StartedAt"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "swaggerDeployment",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "DeploymentListResponse": {
      "type": "object",
      "title": "DeploymentListResponse is the paginated envelope for ListDeployments.",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DeploymentListItem"
          },
          "x-go-name": "Items"
        },
        "limit": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Limit"
        },
        "offset": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Offset"
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-name": "swaggerDeploymentListResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "ServerError": {
      "type": "object",
      "title": "Server error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerServerError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "UnauthorizedError": {
      "type": "object",
      "title": "Unauthorized error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerUnauthorizedError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}